Renamed generated schema func to UseSchema.
